Notebooks/requirements.txt
## Requirements
pandas: 2.2.2
numpy: 1.26.4
seaborn: 0.13.2
matplotlib: 3.8.4
scipy: 1.13.1
statsmodels: 0.14.2
plotly.express:
 